pyproject.toml
[project]
name = "zero-shot-classifier"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "fastapi>=0.115.12",
    "fpdf2>=2.8.3",
    "plotly>=6.1.2",
    "polars>=1.30.0",
    "streamlit>=1.45.1",
    "torch>=2.7.1",
    "transformers>=4.52.4",
    "uvicorn>=0.34.3",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=8.4.0",
]
